Abstract out root navigator
