+++ umbrella_RN_frontend/components/calendar_screen.js
@@ -1,0 +1,23 @@
+import React from 'react';
+import { StyleSheet, View, Text, Button } from 'react-native';
+import { TabNavigator } from 'react-navigation';
+
+class CalendarScreen extends React.Component {
+  render() {
+    return(
+      <View>
+        <Text style={styles.calendar}>
+           You are now on the calendar view page.
+        </Text>
+      </View>
+    );
+  }
+}
+
+export default CalendarScreen;
+
+export const styles = StyleSheet.create({
+  calendar: {
+    margin: 50,
+  },
+});
