+++ umbrella_RN_frontend/navigation/root_navigator.js
@@ -1,29 +1,34 @@
-import React from 'react';
+import React, { Component } from 'react';
 import { StyleSheet, View, Text, Button } from 'react-native';
 import { TabNavigator } from 'react-navigation';
-import LoginForm from './login_form';
-import ProfileScreen from '../components/profile_page';
-import FeedScreen from '../components/feed';
+import LoginForm from '../components/login_form';
+import ProfileScreen from '../components/profile_screen';
+import FeedScreen from '../components/feed_screen';
+import CalendarScreen from '../components/calendar_screen';
+import MessageScreen from '../components/message_screen';
+
 
 const RootNavigator = TabNavigator({
-  Profile: {
-    screen: ProfileScreen,
-    navigationOptions: {
-      headerTitle: 'Profile'
-    }
+  CalendarScreen: {
+    screen: CalendarScreen,
   },
-  Login: {
-    screen: LoginForm,
-    navigationOptions: {
-      headerTitle: 'Login'
-    }
-  },
-  Feed: {
+  FeedScreen: {
     screen: FeedScreen,
-    navigationOptions: {
-      headerTitle: 'Feed'
-    }
-  }
+  },
+  MessageScreen: {
+    screen: MessageScreen,
+  },
+  ProfileScreen: {
+    screen: ProfileScreen,
+  },
+}, {
+  tabBarPosition: 'bottom',
+  animationEnabled: true,
+  tabBarOptions: {
+    activeTintColor: '#e91e63',
+  },
 });
 
-export const RootNavigator;
+export default RootNavigator;
+
+
